requirements.txt
streamlit
python-dotenv
langchain
langgraph
typing_extensions
ipython
google-generativeai